Meta-tabs visibility fix
The meta tabs that are shown in the playlist view will now no longer
scroll out of view when scrolling the playlist. To this end some minor
changes were made to their css and the ScrollToIndex JavaScript function
(which needed the offset value changed to account for the CSS change).
Additionally, in single.php some <? tags were replaced by <?php in order
to get Kohana to actually find the PHP code valid.